Before coding the fillTables() function inside Selection.swift.
The TableView has a specific format for all TableViewCell which includes three items to be filled: image, type, and plate. These values will be filled with the returned values from the URLRequest. It is yet unknown if the response will return the url of an image, or the image itself, which would be inefficient for the Mobile Data Provider.

As of yet, this method will fill the TableView with hardcoded images, type and plates for the presentation. The server is yet to be provided by the client.